+++ cmd/thanos/receive.go
@@ -5,8 +5,8 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"io/ioutil"
+	"net"
 	"os"
 	"path"
 	"strings"
@@ -140,7 +140,7 @@ func registerReceive(app *extkingpin.App) {
 			}
 			parts := strings.Split(*grpcBindAddr, ":")
 			port := parts[len(parts)-1]
-			*localEndpoint = fmt.Sprintf("%s:%s", hostname, port)
+			*localEndpoint = net.JoinHostPort(hostname, port)
 		}
 
 		return runReceive(
@@ -652,7 +652,7 @@ func migrateLegacyStorage(logger log.Logger, dataDir, defaultTenantID string) er
 		return errors.Wrapf(err, "read legacy data dir: %v", dataDir)
 	}
 
-	if err := os.MkdirAll(defaultTenantDataDir, 0777); err != nil {
+	if err := os.MkdirAll(defaultTenantDataDir, 0750); err != nil {
 		return errors.Wrapf(err, "create default tenant data dir: %v", defaultTenantDataDir)
 	}