+++ pkg/query/storeset.go
@@ -698,18 +698,21 @@ func (s *StoreSet) GetMetadataClients() []metadatapb.MetadataClient {
 	return metadataClients
 }
 
-// GetExemplarsClients returns a list of all active exemplars clients.
-func (s *StoreSet) GetExemplarsClients() []exemplarspb.ExemplarsClient {
+// GetExemplarsStores returns a list of all active exemplars stores.
+func (s *StoreSet) GetExemplarsStores() []*exemplarspb.ExemplarStore {
 	s.storesMtx.RLock()
 	defer s.storesMtx.RUnlock()
 
-	exemplars := make([]exemplarspb.ExemplarsClient, 0, len(s.stores))
+	exemplarStores := make([]*exemplarspb.ExemplarStore, 0, len(s.stores))
 	for _, st := range s.stores {
 		if st.HasExemplarsAPI() {
-			exemplars = append(exemplars, st.exemplar)
+			exemplarStores = append(exemplarStores, &exemplarspb.ExemplarStore{
+				ExemplarsClient: st.exemplar,
+				LabelSets:       st.labelSets,
+			})
 		}
 	}
-	return exemplars
+	return exemplarStores
 }
 
 func (s *StoreSet) Close() {